# Optimize getPodServices