This is a hot path in our code on some occasions. This improves the performance substantially by using simpler check of the map labels.

```
name               old time/op    new time/op    delta
GetPodServices-48    35.3µs ± 1%    10.0µs ± 1%  -71.66%  (p=0.000 n=9+9)

name               old alloc/op   new alloc/op   delta
GetPodServices-48    21.4kB ± 0%     4.2kB ± 0%  -80.63%  (p=0.000 n=10+10)

name               old allocs/op  new allocs/op  delta
GetPodServices-48       468 ± 0%        18 ± 0%  -96.15%  (p=0.000 n=10+10)
```

Behavior is identical, this is purely an optimization

**Please provide a description of this PR:**