@istio-testing 2022-09-22
@howardjohn: The following tests **failed**, say `/retest` to rerun all failed tests or `/retest-required` to rerun all mandatory failed tests:

Test name | Commit | Details | Required | Rerun command
--- | --- | --- | --- | ---
integ-pilot-istiodremote-mc_istio | 4c045ba3b5cc84650fd69710f9e14db3e0df0b46 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/41095/integ-pilot-istiodremote-mc_istio/1572878912214536192) | unknown | `/test integ-pilot-istiodremote-mc`
integ-distroless_istio | 4c045ba3b5cc84650fd69710f9e14db3e0df0b46 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/41095/integ-distroless_istio/1572878907433029632) | unknown | `/test integ-distroless`

<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->